pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nano-memory"
version = "0.1.0"
description = "Persistent memory layer for AI agents — long-term brain across sessions"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2",
    "pyyaml",
    "typer",
    "rich",
    "numpy",
    "sentence-transformers",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
dev = ["pytest", "ruff"]

[project.scripts]
nano-mem = "nano_memory.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["nano_memory*"]

[tool.ruff]
line-length = 100
target-version = "py310"
